requirements.txt
openai
python-dotenv
tzdata
gradio
flask
gunicorn
pyprojroot
PyYAML